add convinience methods

+++ node_function.c
@@ -14,7 +14,7 @@ Data *node_function_evaluate(Node *node, map_t context) {
     LURE_ASSERT(node->list != NULL, "function parameter list must not be NULL");
     int n_params = node->list->n_all;
     Data **params = (Data **)calloc(n_params, sizeof(Data *));
-    for(size_t i = 0; i < n_params; i++) {
+    for(size_t i = 0; i < (size_t)n_params; i++) {
         Node *pnode = node->list->all[i];
         params[i] = pnode->evaluate(pnode, context);
     }
@@ -23,7 +23,7 @@ Data *node_function_evaluate(Node *node, map_t context) {
     Data *ret = function_derive(fname, params, n_params);
     leftRes->clean(leftRes);
     free(leftRes);
-    for(size_t i = 0; i < n_params; i++) {
+    for(size_t i = 0; i < (size_t)n_params; i++) {
         params[i]->clean(params[i]);
         free(params[i]);
     }
